Исправлены ошибки в упражнениях 1, 2, 3, 4

+++ lesson_1/task_1.rb
@@ -1,14 +1,14 @@
 puts "Введите Ваше имя"
 name = gets.chomp.to_s
 puts "Введите Ваш рост"
-increase = gets.chomp.to_i
+increase = gets.chomp.to_f
 
 res = increase - 110
 
 if res >= 0
-	puts "#{name}, Ваш идеальный вес - #{res} кг"
+  puts "#{name}, Ваш идеальный вес - #{res} кг"
 else
-	puts "#{name}, Ваш вес уже оптимальный"
+  puts "#{name}, Ваш вес уже оптимальный"
 end
